Update: expose Theme type and createTheme utility
