+++ src/components/Badge/Badge.ts
@@ -1,4 +1,4 @@
-import { useTheme } from '../Theme/Theme'
+import { useTheme } from '../ComponentryProvider/ComponentryProvider'
 import { element } from '../../utils/element-creator'
 import { type ComponentBaseProps } from '../../utils/base-types'
