+++ src/components/Block/Block.ts
@@ -1,5 +1,5 @@
 import { type FC } from 'react'
-import { useTheme } from '../Theme/Theme'
+import { useTheme } from '../ComponentryProvider/ComponentryProvider'
 import { type ComponentBaseProps } from '../../utils/base-types'
 import { element } from '../../utils/element-creator'
