+++ src/components/ComponentryProvider/ComponentryProvider.stories.mdx
@@ -2,8 +2,8 @@ import { Meta } from '@storybook/addon-docs'
 import { BADGE } from '@geometricpanda/storybook-addon-badges'
 
 import { Button } from '../Button/Button'
-import { Theme } from './Theme'
-import { ThemeConsumer } from './storybook/ThemeConsumer'
+import { ComponentryProvider } from './ComponentryProvider'
+import { ProviderConsumer } from './storybook/ProviderConsumer'
 
 <Meta
   title='Components/Theme'
@@ -13,7 +13,7 @@ import { ThemeConsumer } from './storybook/ThemeConsumer'
 
 # Theme
 
-<Theme
+<ComponentryProvider
   theme={{
     Button: {
       color: 'success',
@@ -23,6 +23,6 @@ import { ThemeConsumer } from './storybook/ThemeConsumer'
 >
   <div>
     <Button>A large successful button!</Button>
-    <ThemeConsumer />
+    <ProviderConsumer />
   </div>
-</Theme>
+</ComponentryProvider>
