+++ src/components/Input/Input.tsx
@@ -1,6 +1,6 @@
 import React, { createContext, useContext, useRef } from 'react'
 import { nanoid } from 'nanoid'
-import { useTheme } from '../Theme/Theme'
+import { useTheme } from '../ComponentryProvider/ComponentryProvider'
 import { type ComponentBaseProps } from '../../utils/base-types'
 import { element } from '../../utils/element-creator'
 import { staticComponent } from '../../utils/static-component-builder'
