+++ src/index.ts
@@ -1,5 +1,8 @@
 // --- Providers
-export { Theme, useTheme } from './components/Theme/Theme'
+export {
+  ComponentryProvider,
+  useTheme,
+} from './components/ComponentryProvider/ComponentryProvider'
 export { Media, useMedia } from './components/Media/Media'
 
 // --- Components
@@ -37,7 +40,8 @@ export { Tooltip } from './components/Tooltip/Tooltip'
 
 // --- Utilities
 export { useActive, useActiveScrollReset, useNoScroll, useVisible } from './hooks'
-export { themeDefaults } from './theme-defaults'
+export { type Theme, createTheme } from './theme/theme'
+export { themeDefaults } from './theme/theme-defaults'
 export { setupOutlineHandlers } from './utils/dom'
 export { borderPlugin } from './utils/tailwind-plugins'
 export {
