+++ src/theme/create-theme.spec.js
@@ -1,5 +1,5 @@
-import { themeDefaults } from '../theme-defaults'
-import { createTheme } from './create-theme'
+import { themeDefaults } from './theme-defaults'
+import { createTheme } from './theme'
 
 describe('merge()', () => {
   it('merges theme values', () => {
